Refactor: Indexer Pages
