+++ client/package.json
@@ -32,8 +32,8 @@
     "react-admin": "^4.9.4",
     "react-admin-lb4": "^1.0.3",
     "react-alice-carousel": "^2.7.1",
-    "react-component-indexer": "^1.0.3",
     "react-apexcharts": "^1.4.0",
+    "react-component-indexer": "^1.0.4",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
     "react-icons": "^4.8.0",
