--- client/src/pages/Dashboard/Administrator/Admin.jsx
@@ -1,95 +1,0 @@
-import React from 'react'
-// import ReactApexChart from 'react-apexcharts'
-
-// const data = [
-//   {
-//     id: 1,
-//     // img: logoCompañia,
-//     profesion: 'Arquitecto',
-//     time: 'Now',
-//     location: 'Peru',
-//     salario: '2000$',
-//     company: 'Jet Constructora',
-//     status: 'disponible',
-//   },
-//   {
-//     id: 2,
-//     // image: logoConstruccion,
-//     profesion: 'Ingeniero',
-//     time: '12Hrs',
-//     location: 'Peru',
-//     salario: '4000$',
-//     company: 'Constructora Cop.',
-//     status: 'no disponible',
-//   },
-//   {
-//     id: 3,
-//     // image: logo3,
-//     profesion: 'Albañil',
-//     time: '24Hrs',
-//     location: 'Peru',
-//     salario: '1500$.',
-//     company: 'Citmat Cop.',
-//     status: 'disponible',
-//   },
-//   {
-//     id: 4,
-//     // image: logo4,
-//     profesion: 'Maestro de Obra',
-//     time: '7Hrs',
-//     location: 'Peru',
-//     salario: '2500$.',
-//     company: 'Soluciones C.O.',
-//     status: 'disponible',
-//   },
-//   {
-//     id: 5,
-//     // image: logo5,
-//     profesion: 'Ingeniero Civil',
-//     time: '10Hrs',
-//     location: 'Peru',
-//     salario: '3500$',
-//     company: 'Obras Sil.',
-//     status: 'disponible',
-//   },
-//   {
-//     id: 6,
-//     // image: logo6,
-//     profesion: 'Arquitecto',
-//     time: '2Hrs',
-//     location: 'Peru',
-//     salario: '1800$.',
-//     company: 'Corporacion A.S.',
-//     status: 'no disponible',
-//   },
-//   {
-//     id: 7,
-//     // image: logo7,
-//     profesion: 'Ayudante de Obra',
-//     time: '4Hrs',
-//     location: 'Peru',
-//     salario: '2000$',
-//     company: 'Contrucciones C.A',
-//     status: 'disponible',
-//   },
-//   {
-//     id: 8,
-//     // image: logo8,
-//     profesion: 'Ingeniero Civil',
-//     time: '1Hrs',
-//     location: 'Peru',
-//     salario: '2000$',
-//     company: 'Civil Cop.',
-//     status: 'disponible',
-//   },
-// ]
-
-const Admin = () => {
-  return (
-    <div>
-      <h1>holaa</h1>
-    </div>
-  )
-}
-
-export default Admin
