+++ client/src/pages/Dashboard/Company/FavedTalents.jsx
