+++ client/src/pages/index.js
@@ -1,3 +1,4 @@
+export { AccountContext } from 'pages/Account/AccountContext'
 export { ButtonContainerLg } from 'pages/Account/ButtonContainerLg'
 export { AppContext } from 'pages/Account/OnBoarding/AppContext'
 export { AppContextProvider } from 'pages/Account/OnBoarding/AppContextProvider'
@@ -11,18 +12,27 @@ export { Step6 } from 'pages/Account/steps/Step6'
 export { AuthLayout } from 'pages/Auth/AuthLayout'
 export { ForgotPass } from 'pages/Auth/ForgotPass'
 export { Login } from 'pages/Auth/Login'
+export { LoginBack } from 'pages/Auth/LoginBack'
 export { Register } from 'pages/Auth/Register'
+export { RegisterBack } from 'pages/Auth/RegisterBack'
 export { Companies } from 'pages/Community/Companies/Companies'
 export { default as Media } from 'pages/Community/Media'
 export { TalentDetail } from 'pages/Community/Talents/TalentDetail'
 export { Talents } from 'pages/Community/Talents/Talents'
 export { AdminDash } from 'pages/Dashboard/Admin/AdminDash'
 export { CompanyDash } from 'pages/Dashboard/Company/CompanyDash'
+export { default as CompanyOffers } from 'pages/Dashboard/Company/CompanyOffers'
+export { default as CreateOffer } from 'pages/Dashboard/Company/CreateOffer'
+export { default as FavedTalents } from 'pages/Dashboard/Company/FavedTalents'
 export { MarketProfessionals } from 'pages/Dashboard/Company/MarketProfessionals'
 export { Dashboard } from 'pages/Dashboard/Dashboard'
 export { DashboardLayout } from 'pages/Dashboard/DashboardLayout'
-export { default as Offers } from 'pages/Dashboard/Professional/Offers'
+export { default as DashCards } from 'pages/Dashboard/MinCards/DashCards'
+export { default as DashCards_v2 } from 'pages/Dashboard/MinCards/DashCards_v2'
+export { default as DashProfessionalOffers } from 'pages/Dashboard/Professional/DashProfessionalOffers'
+export { DashUserProfile } from 'pages/Dashboard/Professional/DashUserProfile'
 export { ProfessionalDash } from 'pages/Dashboard/Professional/ProfessionalDash'
+export { SavedProfessionalOffers } from 'pages/Dashboard/Professional/SavedProfessionalOffers'
 export { Error404 } from 'pages/Errors/Error404'
 export { AddInputDynamic } from 'pages/Form/AddInputDynamic'
 export { CVForm } from 'pages/Form/CVForm'
@@ -34,12 +44,16 @@ export { default as UserProfile } from 'pages/Form/UserProfile'
 export { UserProfileForm } from 'pages/Form/UserProfileForm'
 export { default as HomeEmpresa } from 'pages/HomeEmpresa'
 export { default as About } from 'pages/Info/About'
+export { default as Admin } from 'pages/Info/Admin'
 export { default as Contact } from 'pages/Info/Contact'
 export { default as InfoProfessionals } from 'pages/Info/InfoProfessionals'
 export { default as News } from 'pages/Info/News'
 export { OfferDetail } from 'pages/Info/OfferDetail'
+export { default as Offers } from 'pages/Info/Offers'
+export { default as Perfil } from 'pages/Info/Perfil'
 export { PricingEmpresa } from 'pages/Info/pricing/PricingEmpresa'
 export { default as PricingPage } from 'pages/Info/pricing/PricingPage'
 export { PricingProfesional } from 'pages/Info/pricing/PricingProfesional'
 export { Into } from 'pages/Into'
 export { Landing } from 'pages/Landing'
+export { default as Footer } from 'pages/footer/Footer'
